Add tests to TS actions
