+++ .github/actions/compute-sha256/jest.setup.js
@@ -1,0 +1,1 @@
+console.error = jest.fn();
