+++ .github/actions/generate-attestations/Makefile
@@ -1,9 +1,9 @@
 SHELL := /bin/bash
-OUTPUT_FORMAT = $(shell if [ "${GITHUB_ACTIONS}" == "true" ]; then echo "github"; else echo ""; fi)
+ACTION_NAME = $(shell basename "$$(pwd)")
 
 .PHONY: help
 help: ## Shows all targets and help from the Makefile (this message).
-	@echo "detect-workflow Makefile"
+	@echo "$(ACTION_NAME) Makefile"
 	@echo "Usage: make [COMMAND]"
 	@echo ""
 	@grep --no-filename -E '^([/a-z.A-Z0-9_%-]+:.*?|)##' $(MAKEFILE_LIST) | \
@@ -37,10 +37,9 @@ clean:
 #####################################################################
 
 .PHONY: unit-test
-unit-test: ## Runs all unit tests.
-	npm run all
+unit-test: node_modules/.installed ## Runs all unit tests.
+	npm run test
 
 .PHONY: lint
 lint: node_modules/.installed ## Runs eslint.
 	npm run lint
-
