+++ .github/actions/generate-attestations/package.json
@@ -2,7 +2,7 @@
   "name": "generate-attestations",
   "version": "0.0.1",
   "description": "",
-  "main": "src/main.ts",
+  "main": "lib/main.js",
   "scripts": {
     "ci": "npm ci",
     "build": "tsc",
