+++ .github/actions/privacy-check/Makefile
@@ -1,9 +1,9 @@
 SHELL := /bin/bash
-OUTPUT_FORMAT = $(shell if [ "${GITHUB_ACTIONS}" == "true" ]; then echo "github"; else echo ""; fi)
+ACTION_NAME = $(shell basename "$$(pwd)")
 
 .PHONY: help
 help: ## Shows all targets and help from the Makefile (this message).
-	@echo "privacy-check Makefile"
+	@echo "$(ACTION_NAME) Makefile"
 	@echo "Usage: make [COMMAND]"
 	@echo ""
 	@grep --no-filename -E '^([/a-z.A-Z0-9_%-]+:.*?|)##' $(MAKEFILE_LIST) | \
@@ -35,6 +35,10 @@ clean:
 
 ## Testing
 #####################################################################
+
+.PHONY: unit-test
+unit-test: node_modules/.installed ## Runs all unit tests.
+	npm run test
 
 .PHONY: lint
 lint: node_modules/.installed ## Runs eslint.
