+++ .github/actions/privacy-check/jest.setup.js
@@ -1,0 +1,1 @@
+console.error = jest.fn();
