+++ .github/actions/sign-attestations/package.json
@@ -10,12 +10,14 @@
     "format-check": "prettier --check '**/*.ts'",
     "package": "ncc build --source-map",
     "lint": "eslint src/**/*.ts",
-    "all": "npm ci && npm run format && npm run lint && npm run build && npm run package"
+    "all": "npm ci && npm run format && npm run lint && npm run build && npm test && npm run package",
+    "test": "jest"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "devDependencies": {
+    "@types/jest": "^29.2.4",
     "@types/node": "18.11.11",
     "@typescript-eslint/eslint-plugin": "5.45.1",
     "@typescript-eslint/parser": "5.45.1",
@@ -23,6 +25,7 @@
     "eslint": "8.29.0",
     "eslint-plugin-github": "4.6.0",
     "prettier": "2.8.0",
+    "ts-jest": "^29.0.3",
     "typescript": "4.9.3"
   },
   "dependencies": {
