+++ .github/actions/verify-token/jest.setup.js
@@ -1,0 +1,1 @@
+console.error = jest.fn();
