+++ Makefile
@@ -21,16 +21,21 @@ help: ## Shows all targets and help from the Makefile (this message).
 #####################################################################
 
 .PHONY: unit-test
-unit-test: 
+unit-test: ## Run Go unit tests.
 	# Run unit tests for the detect-workflow action.
 	make -C .github/actions/detect-workflow/ unit-test
 	go mod vendor
 	go test -mod=vendor -v ./...
 
 .PHONY: ts-test
-ts-test: 
+ts-test: ## Run TypeScript unit tests.
 	# Run unit tests for the generate-attestations action.
-	make -C .github/actions/generate-attestations/ unit-test
+	make -C .github/actions/compute-sha256 unit-test
+	make -C .github/actions/generate-attestations unit-test
+	make -C .github/actions/privacy-check unit-test
+	make -C .github/actions/sign-attestations unit-test
+	make -C .github/actions/verify-token unit-test
+	make -C actions/setup-token unit-test
 
 
 ## Linters
@@ -80,8 +85,11 @@ shellcheck: ## Runs the shellcheck linter.
 .PHONY: eslint
 eslint: ## Runs the eslint linter.
 	make -C .github/actions/compute-sha256 lint
-	make -C .github/actions/privacy-check lint
 	make -C .github/actions/generate-attestations lint
+	make -C .github/actions/privacy-check lint
+	make -C .github/actions/sign-attestations lint
+	make -C .github/actions/verify-token lint
+	make -C actions/setup-token lint
 
 .PHONY: yamllint
 yamllint: ## Runs the yamllint linter.
